refactor: rename files in the src/Terminal/Window directory
- window.hpp to Window.hpp
- window.cpp to Window.cpp
- window_size.hpp to WindowSize.hpp
