+++ tests/Editor/Editor.test.cpp
@@ -26,7 +26,7 @@
 #include "Cursor/Cursor.hpp"
 #include "Offset/Offset.hpp"
 #include "ScreenBuffer/ScreenBuffer.hpp"
-#include "Terminal/Window/window.hpp"
+#include "Terminal/Window/Window.hpp"
 #include "Utilities/Utilities.hpp"
 #include "gmock/gmock.h"
 #include <cstring>
