+++ tests/Window/Window.test.cpp
@@ -1,4 +1,4 @@
-#include "Terminal/Window/window.hpp"
+#include "Terminal/Window/Window.hpp"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
